Story 4: refactor — align FE acceptance-test infra to "register" naming
Completes the ubiquitous-language rename by extending it from production
(previous commit) into the black-box acceptance test infra. Pure symbol/file
rename, no behavior change; the URL /api/admin/users stays.

- create-user-backend.statements.ts -> register-user-backend.statements.ts
  (CreateUserBackendStatements -> RegisterUserBackendStatements; givenCreateUser*
  -> givenRegisterUser*; releaseCreateUser/routeCreateUser -> releaseRegisterUser/
  routeRegisterUser)
- support/create-user-fixture.ts -> support/register-user-fixture.ts (import paths
  updated in admin-users-backend, register-user-modal.statements, users-page.statements)
- real-auth-backend.statements.ts: createUserAsAdmin -> registerUserAsAdmin (+ caller
  in account-lifecycle.fullstack.spec.ts); comments/strings aligned
- register-user-modal.spec.ts: createUserBackend var + given* calls + comment aligned

Left intentionally: HTTP-status-derived names (fulfillCreated, releaseInFlightCreate)
are not the domain term. No create-user identifier remains in frontend/acceptance.
Full frontend lint green (oxlint + eslint + prettier + vue-tsc).
