+++ frontend/acceptance/tests/statements/support/register-user-fixture.ts
